+++ extensions/xnat/src/elements/index.js
@@ -1,5 +1,6 @@
 // ICR/XNAT icons
 import xnatOhifLogo from './icons/xnat-ohif-logo.svg';
+import xnatOhifLogoAnim from './icons/xnat-ohif-logo-animated.svg';
 import xnatIcrLogo from './icons/xnat-icr-logo.svg';
 import xnatSettings from './icons/xnat-settings.svg';
 import xnatAnnotations from './icons/xnat-annotations.svg';
@@ -45,6 +46,7 @@ import xnatSync from './icons/xnat-sync.svg';
 import xnatAnnotate from './icons/xnat-annotate.svg';
 import xnatViewportOverlay from './icons/xnat-viewport-overlay.svg';
 import xnatStack from './icons/xnat-stack.svg';
+import xnatScanGroup from './icons/xnat-scan-group.svg';
 
 import Loader from './Loader/Loader';
 
@@ -60,6 +62,7 @@ import CheckIcon from './CheckIcon/CheckIcon';
 
 const XNATICONS = {
   'xnat-ohif-logo': xnatOhifLogo,
+  'xnat-ohif-logo-anim': xnatOhifLogoAnim,
   'xnat-icr-logo': xnatIcrLogo,
   'xnat-settings': xnatSettings,
   'xnat-annotations': xnatAnnotations,
@@ -105,6 +108,7 @@ const XNATICONS = {
   'xnat-annotate': xnatAnnotate,
   'xnat-viewport-overlay': xnatViewportOverlay,
   'xnat-stack': xnatStack,
+  'xnat-scan-group': xnatScanGroup,
 };
 
 export {