+++ extensions/xnat/src/utils/IO/classes/_tempDCMJSFork/copiedDCMJSInternals/utilities/orientation/nearlyEqual.js
@@ -19,10 +19,10 @@ export default function nearlyEqual(a, b, epsilon) {
   const absB = Math.abs(b);
   const diff = Math.abs(a - b);
 
-  if (a == b) {
+  if (a === b) {
     // Shortcut, handles infinities
     return true;
-  } else if (a == 0 || b == 0 || absA + absB < Number.EPSILON) {
+  } else if (a === 0 || b === 0 || absA + absB < epsilon * epsilon) {
     // A or b is zero or both are extremely close to it
     // relative error is less meaningful here
     return diff < epsilon;