+++ platform/viewer/src/routes/NotFound.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import { Icon } from '@ohif/ui';
 import './NotFound.css';
 import { Link } from 'react-router-dom';
 import { useAppContext } from '../context/AppContext';
@@ -9,6 +10,13 @@ export default function NotFound({ message = 'Sorry, this page does not exist.',
 
   return (
     <div className={'not-found'}>
+      <div>
+        {showGoBackButton ? (
+          <Icon name="xnat-ohif-logo" className="xnat-logo" />
+        ) : (
+          <Icon name="xnat-ohif-logo-anim" className="xnat-logo" />
+        )}
+      </div>
       <div>
         <h4>{message}</h4>
         {/*{showGoBackButton && context.appConfig.showStudyList && (*/}