+++ platform/viewer/src/routes/XNATStandaloneRouting.js
@@ -147,13 +147,12 @@ class XNATStandaloneRouting extends Component {
           //   );
           // }
 
-          const data = JSON.parse(jsonString);
-          log.info(data);
+          // log.info(JSON.stringify(jsonString, null, 2));
 
-          let studies = data.studies;
+          const data = JSON.parse(jsonString);
+
           const dicomWebParameters = [];
           const studyInstanceUIDs = [];
-
           if (data.isDicomWeb) {
             // Possibly a DICOMweb study
             dicomWebParameters.push({
@@ -172,7 +171,15 @@ class XNATStandaloneRouting extends Component {
             });
           }
 
-          studies[0].StudyDescription = experimentLabel || experimentId;
+          const studies = data.studies.filter(
+            study => study.series !== undefined
+          );
+
+          data.studies = studies;
+          if (studies.length === 0) {
+            resolve({ studies: [], studyInstanceUIDs: [] });
+            return;
+          }
 
           commandsManager.runCommand('xnatSetSession', {
             json: data,
@@ -185,7 +192,15 @@ class XNATStandaloneRouting extends Component {
             },
           });
 
-          resolve({ studies, studyInstanceUIDs, dicomWebParameters });
+          data.studies[0].StudyDescription = experimentLabel || experimentId;
+
+          console.log(data);
+
+          resolve({
+            studies: data.studies,
+            studyInstanceUIDs: [],
+            dicomWebParameters,
+          });
         });
 
         // Open the Request to the server for the JSON data
@@ -238,7 +253,9 @@ class XNATStandaloneRouting extends Component {
 
             for (let i = 0; i < jsonFiles.length; i++) {
               const experimentJsonI = jsonFiles[i];
-              const studiesI = experimentJsonI.studies;
+              const studiesI = experimentJsonI.studies.filter(
+                study => study.series !== undefined
+              );
 
               // Exclude studies with no instances
               if (studiesI.length === 0) {
@@ -583,13 +600,13 @@ async function updateMetaDataProvider(studies) {
     StudyInstanceUID = study.StudyInstanceUID;
     for (let series of study.series) {
       SeriesInstanceUID = series.SeriesInstanceUID;
-      const { is4D, numberOfSubInstances } = metadataUtils.isDataset4D(
-        series.instances
-      );
-      series.is4D = is4D;
-      series.numberOfSubInstances = numberOfSubInstances;
+      const series4DConfig = metadataUtils.isDataset4D(series.instances);
+      series._4DConfig = series4DConfig;
+      // Check if the series is 4D multi-stack
       series.isMultiStack =
-        is4D && metadataUtils.isSameOrientation(series.instances);
+        series4DConfig.isValid4D &&
+        metadataUtils.isSameOrientation(series.instances);
+
       await Promise.all(
         series.instances.map(async (instance, instanceIndex) => {
           const { url: imageId, metadata: naturalizedDicom } = instance;
@@ -611,33 +628,66 @@ async function updateMetaDataProvider(studies) {
           // PaletteColorLookupTableData is loaded conditionally in metadataProvider.addInstance
           // OverlayData is loaded conditionally in metadataProvider.addInstance
 
+          const isEnhancedSOP = metadataUtils.isEnhancedSOP(
+            naturalizedDicom.SOPClassUID
+          );
+          let shouldFetchDataset = false;
+          if (
+            naturalizedDicom.NumberOfFrames === 1 &&
+            naturalizedDicom.PhotometricInterpretation === 'PALETTE COLOR'
+          ) {
+            shouldFetchDataset = true;
+          } else if (
+            series.isMultiStack &&
+            series4DConfig.sameIppIndices.includes(instanceIndex)
+          ) {
+            shouldFetchDataset = true;
+          } else if (isEnhancedSOP && !series4DConfig.hasMultiFrameInstances) {
+            shouldFetchDataset = true;
+          }
+
           // Add instance to metadata provider.
           const addedInstance = await metadataProvider.addInstance(
             naturalizedDicom,
             {
               imageId,
-              shouldFetchDataset: is4D && instanceIndex < numberOfSubInstances,
+              shouldFetchDataset,
             }
           );
-          if (metadataUtils.isEnhancedSOP(addedInstance.SOPClassUID)) {
-            const naturalizedMetadataList = metadataUtils.parseEnhancedSOP(
-              addedInstance
-            );
-            if (naturalizedMetadataList && naturalizedMetadataList.length > 0) {
-              const subInstances = [];
-              for (let j = 0; j < naturalizedMetadataList.length; j++) {
-                subInstances.push({
-                  metadata: naturalizedMetadataList[j],
-                  url: `${imageId}?frame=${j}`,
-                });
-              }
-              series.isEnhanced = metadataUtils.isSameOrientation(subInstances);
-              series.subInstances = subInstances;
+
+          if (isEnhancedSOP) {
+            // NM modality
+            if (addedInstance.SOPClassUID === '1.2.840.10008.5.1.4.1.1.20') {
+              series.isEnhanced = true;
+            } else if (series4DConfig.hasMultiFrameInstances) {
+              // ToDo: should we add further flags here?
+            } else {
+              // Create individual frame metadata from the enhanced instance
+              const naturalizedMetadataList = metadataUtils.parseEnhancedSOP(
+                addedInstance
+              );
               if (
-                series.isEnhanced &&
-                metadataUtils.isDataset4D(subInstances).is4D
+                naturalizedMetadataList &&
+                naturalizedMetadataList.length > 0
               ) {
-                series.isMultiStack = true;
+                const subInstances = [];
+                for (let j = 0; j < naturalizedMetadataList.length; j++) {
+                  subInstances.push({
+                    metadata: naturalizedMetadataList[j],
+                    url: `${imageId}?frame=${j}`,
+                  });
+                }
+                series.isEnhanced = isEnhancedSOP;
+                const isUniformOrientation = metadataUtils.isSameOrientation(
+                  subInstances
+                );
+                // Update the series 4D config from sub-instances
+                const series4DConfig = metadataUtils.isDataset4D(subInstances);
+                series._4DConfig = series4DConfig;
+                series.subInstances = subInstances;
+                if (isUniformOrientation && series4DConfig.isValid4D) {
+                  series.isMultiStack = true;
+                }
               }
             }
           }
@@ -812,12 +862,23 @@ function isValidOverlayDisplaySet(displaySetI, refIop, displaySetJ) {
 function getImageOrientationPatient(displaySet) {
   const firstImage = displaySet.images[0];
   let imageId = firstImage.getData().url;
-  if (displaySet.isMultiFrame) {
-    imageId += '?frame=0';
+
+  if (displaySet.Modality === 'NM') {
+    const instance = cornerstone.metaData.get('instance', imageId);
+    if (instance) {
+      const {
+        ImageOrientationPatient,
+      } = metadataUtils.getImagePlaneInformation(instance, 0);
+      return ImageOrientationPatient;
+    }
+  } else {
+    if (displaySet.isMultiFrame) {
+      imageId += '?frame=0';
+    }
+    const { imageOrientationPatient } = cornerstone.metaData.get(
+      'imagePlaneModule',
+      imageId
+    );
+    return imageOrientationPatient;
   }
-  const { imageOrientationPatient } = cornerstone.metaData.get(
-    'imagePlaneModule',
-    imageId
-  );
-  return imageOrientationPatient;
 }